// despachodelivery.client/src/app/shared/interfaces/facturaklkheader.interface.ts
export interface FacturaHeadKlk {
    numFactura:    string;
    fechaFactura:  Date;
    codCliente:    string;
    nomCliente:    string;
    serial:        string;
    codArticulo:   string;
    descripcion:   string;
    precioUSD:     number;
    precioPoliza:  number;
    codigoAlmacen: string;
    articulo:      string;
}